+++ frontend/components/search/Timeline/Timeline.tsx
@@ -1,35 +1,13 @@
+import TimelineMark, {
+  Mark,
+  TimelineMarkModule,
+  TimelineMarkProps,
+} from "@/components/search/Timeline/TimelineMark";
 import { SerpModule } from "@/types/modules";
-import Compress from "@mui/icons-material/Compress";
-import {
-  Box,
-  Slider as MuiSlider,
-  sliderClasses,
-  SliderMark as MuiSliderMark,
-  SliderProps,
-  SliderThumb,
-  styled,
-  Tooltip,
-  TooltipProps,
-} from "@mui/material";
-import {
-  ComponentProps,
-  Dispatch,
-  FC,
-  forwardRef,
-  memo,
-  MouseEventHandler,
-  ReactElement,
-  ReactNode,
-  RefObject,
-  SetStateAction,
-  TouchEventHandler,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { Box, Slider as MuiSlider, sliderClasses, SliderProps, styled } from "@mui/material";
+import { FC, TouchEventHandler, useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { throttle } from "throttle-debounce";
+import TimelineThumb, { TimelineThumbProps } from "./TimelineThumb";
 
 const Slider: FC<SliderProps<"span", { componentsProps?: { mark: TimelineMarkProps } }>> = styled(
   MuiSlider
@@ -45,12 +23,6 @@ const Slider: FC<SliderProps<"span", { componentsProps?: { mark: TimelineMarkPro
   },
 });
 
-const BreakIcon = styled(Compress)({
-  transform: "translate(-25%, -50%)",
-  color: "gray",
-  backgroundColor: "white",
-});
-
 type PositionedModule<T extends Partial<SerpModule>> = T & Required<Pick<T, "timelinePosition">>;
 
 function positionedModuleTypeGuard<T extends Partial<SerpModule>>(
@@ -59,159 +31,6 @@ function positionedModuleTypeGuard<T extends Partial<SerpModule>>(
   return module.timelinePosition != null;
 }
 
-type TimelineMarkModule<T extends HTMLElement = HTMLElement> = Required<
-  Pick<SerpModule, "timelinePosition" | "title" | "absoluteUrl">
-> & { ref: RefObject<T> };
-
-interface TimelineMarkProps extends ComponentProps<typeof MuiSliderMark> {
-  viewStateRegistry: Map<string, Dispatch<SetStateAction<boolean>>>;
-  marks: Mark[];
-}
-
-const TimelineModuleMark: FC<
-  Pick<TimelineMarkBaseProps, "viewStateRegistry" | "muiMarkProps"> & {
-    mark: Mark & { isBreak: false };
-  }
-> = ({ mark, viewStateRegistry, muiMarkProps }) => (
-  <TimelineMarkBase
-    tooltipTitle={mark.module.title}
-    markBoxProps={{
-      position: "relative",
-      bottom: "3px",
-      padding: "5px",
-    }}
-    onClick={() => {
-      const moduleCard = mark.module.ref.current;
-      moduleCard?.scrollIntoView({ behavior: "smooth" });
-      moduleCard?.click();
-    }}
-    registryKey={mark.module.absoluteUrl}
-    mark={mark}
-    viewStateRegistry={viewStateRegistry}
-    muiMarkProps={muiMarkProps}
-  />
-);
-
-const TimelineBreakMark: FC<
-  Pick<TimelineMarkBaseProps, "viewStateRegistry" | "muiMarkProps"> & {
-    mark: Mark & { isBreak: true };
-  }
-> = ({ mark, viewStateRegistry, muiMarkProps }) => (
-  <TimelineMarkBase
-    tooltipTitle={mark.label}
-    tooltipProps={{ zIndex: 1 }}
-    markBoxProps={{ children: <BreakIcon /> }}
-    registryKey={String(mark.value)}
-    mark={mark}
-    viewStateRegistry={viewStateRegistry}
-    muiMarkProps={muiMarkProps}
-  />
-);
-
-interface TimelineMarkBaseProps {
-  tooltipTitle: ReactNode;
-  tooltipProps?: NonNullable<TooltipProps["componentsProps"]>["tooltip"];
-  markBoxProps: ComponentProps<typeof Box>;
-  onClick?: MouseEventHandler;
-  registryKey: string;
-  mark: Mark;
-  viewStateRegistry: Map<string, Dispatch<SetStateAction<boolean>>>;
-  muiMarkProps: ComponentProps<typeof MuiSliderMark>;
-}
-
-const TimelineMarkBase: FC<TimelineMarkBaseProps> = ({
-  tooltipTitle,
-  tooltipProps,
-  markBoxProps,
-  onClick,
-  registryKey,
-  mark,
-  viewStateRegistry,
-  muiMarkProps,
-}) => {
-  const [tooltipOpen, setTooltipOpen] = useState(false);
-  const [inView, setInView] = useState(
-    mark.isBreak ? false : isElementInViewport(mark.module.ref.current)
-  );
-
-  viewStateRegistry.set(registryKey, setInView);
-  return (
-    <Tooltip
-      title={
-        <Box whiteSpace={"nowrap"} data-testid={"timelineTooltip"}>
-          {tooltipTitle}
-        </Box>
-      }
-      arrow
-      placement={"right"}
-      open={inView || tooltipOpen}
-      onOpen={() => setTooltipOpen(true)}
-      onClose={() => setTooltipOpen(false)}
-      // MUI mis-types "popper" key as PopperProps instead of Partial<PopperProps>
-      componentsProps={{
-        popper: { disablePortal: true } as any,
-        tooltip: { onClick, ...tooltipProps },
-      }}
-    >
-      <MuiSliderMark {...muiMarkProps} data-testid={"timelineBreak"}>
-        <Box
-          ref={(ref: HTMLSpanElement) => {
-            mark.ref = ref;
-          }}
-          onClick={onClick}
-          {...markBoxProps}
-        />
-      </MuiSliderMark>
-    </Tooltip>
-  );
-};
-
-const TimelineMark: FC<TimelineMarkProps> = memo(
-  function TimelineMark({ marks, viewStateRegistry, ...muiMarkProps }) {
-    // Mui passes this prop but doesn't type it.
-    const mark = marks[(muiMarkProps as unknown as { "data-index": number })["data-index"]];
-    return mark.isBreak ? (
-      <TimelineBreakMark
-        mark={mark}
-        viewStateRegistry={viewStateRegistry}
-        muiMarkProps={muiMarkProps}
-      />
-    ) : (
-      <TimelineModuleMark
-        mark={mark}
-        viewStateRegistry={viewStateRegistry}
-        muiMarkProps={muiMarkProps}
-      />
-    );
-  },
-  (prevProps, nextProps) => {
-    const keys: Array<keyof TimelineMarkProps> = ["marks", "viewStateRegistry"];
-    for (const key of keys) {
-      if (prevProps[key] !== nextProps[key]) return false;
-    }
-    return true;
-  }
-);
-
-interface TimelineThumbProps extends ComponentProps<typeof SliderThumb> {
-  refs: HTMLSpanElement[];
-  "data-index": number;
-}
-
-const TimelineThumb: FC<TimelineThumbProps> = ({ refs, ...props }) => {
-  // SliderThumb does forward its ref, but it is mis-typed, so we have to cast it.
-  const SliderThumbWithRef = SliderThumb as ReturnType<typeof forwardRef>;
-  return (
-    <SliderThumbWithRef
-      {...props}
-      ref={(ref: HTMLSpanElement) => {
-        refs[props["data-index"]] = ref;
-      }}
-      data-testid={"timelineThumb"}
-    />
-  );
-};
-
 export type TimelineProps<T extends HTMLElement = HTMLElement> = {
   modules: Array<
     Omit<TimelineMarkModule<T>, "timelinePosition"> &
@@ -219,17 +38,6 @@ export type TimelineProps<T extends HTMLElement = HTMLElement> = {
   >;
 } & Pick<TimelineMarkProps, "viewStateRegistry"> &
   Partial<Pick<SliderProps, "sx">>;
-
-type Mark<T extends HTMLElement = HTMLElement> = Exclude<
-  SliderProps["marks"],
-  boolean | undefined
->[number] &
-  (
-    | { isBreak: false; module: TimelineMarkModule<T> }
-    | { isBreak: true; label: ReactElement; module?: never }
-  ) & {
-    ref?: HTMLSpanElement;
-  };
 
 interface TimelineCalculations extends Required<Pick<SliderProps, "min" | "max">> {
   marks: Mark[];
@@ -457,15 +265,4 @@ function formatYbp(ybp: number, thisYear: number) {
     ? year.toPrecision(3)
     : Math.round(year).toString().replace(commaRegex, "$1,");
   return `${yearStr}${multiplier} ${type}`;
-}
-
-function isElementInViewport(el: HTMLElement | null) {
-  if (el === null) return false;
-  const { top, bottom, left, right } = el.getBoundingClientRect();
-  return (
-    top < (window.innerHeight || document.documentElement.clientHeight) &&
-    bottom > 0 &&
-    left < (window.innerWidth || document.documentElement.clientWidth) &&
-    right > 0
-  );
 }
