+++ frontend/components/search/Timeline/index.ts
@@ -1,0 +1,4 @@
+import Timeline, { TimelineProps } from "./Timeline";
+
+export default Timeline;
+export type { TimelineProps };
